+++ src/core/baseClass/Epic.js
@@ -1,4 +1,8 @@
 export default class Epic {
+    constructor({ $http }) {
+        this.$http = $http
+    }
+
     // 类型是epic
     $type = 'epic';
 }