+++ src/core/inject/injectHttp.js
@@ -1,0 +1,5 @@
+export default function injectHttp(http) {
+    return function (Target) {
+        return new Target({ $http: http })
+    }
+}