+++ src/page/room/epics/demo.epic.js
@@ -1,7 +1,6 @@
 import {
     mergeMap, map,
 } from 'rxjs/operators';
-import http$ from '@home/util/http';
 import Epic from '@home/core/baseClass/Epic';
 
 export default class ListEpic extends Epic {
@@ -10,12 +9,17 @@ export default class ListEpic extends Epic {
 
     fetchList(action$) {
         return action$.pipe(
-            mergeMap(({ payload }) => http$.getJSON(`https://api.github.com/users/${payload.id}?param=${payload.id}`).pipe(
-                map((response) => ({
-                    type: 'list/showList',
-                    payload: response,
-                })),
-            )),
+            mergeMap(({ payload }) => {
+                return this.$http
+                    .$send({
+                        url: `https://api.github.com/users/${payload.id}?param=${payload.id}`
+                    }).pipe(
+                        map((response) => ({
+                            type: 'list/showList',
+                            payload: response,
+                        })),
+                    )
+            }),
         );
     }
 }