+++ src/page/room/epics/index.js
@@ -1,6 +1,9 @@
-import combineEpics from '@home/core/combineEpics';
+// import combineEpics from '@home/core/combineEpics';
 
 import demo from './demo.epic';
 import demo2 from './demo2.epic';
 
-export default combineEpics(demo, demo2);
+export default [
+    demo,
+    demo2,
+];